requirements.txt
torch==1.11.0
torchvision==0.12.0
absl-py==1.0.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
astunparse==1.6.3
black==21.4b2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
Flask==2.1.2
flatbuffers==2.0
fonttools==4.33.3
future==0.18.2
fvcore==0.1.5.post20220504
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.0
h5py==3.6.0
hydra-core==1.1.2
idna==3.3
importlib-metadata==4.11.3
iopath==0.1.9
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.2
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
omegaconf==2.1.2
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
portalocker==2.4.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pydot==1.4.2
pyparsing==3.0.8
python-dateutil==2.8.2
pywin32==225
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2

tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wincertstore==0.2
wrapt==1.14.1
yacs==0.1.8
zipp==3.8.0
